requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
pyyaml>=6.0
tqdm>=4.65.0
ultralytics>=8.3.0
pycocotools>=2.0.7
supervision>=0.21.0
matplotlib>=3.8.0
huggingface_hub>=0.20.0
lightly-train>=0.11.2
jupyter>=1.0.0
